Optimization and uppload of validation data to git

+++ src/pacmap_pure_cpp/pacmap_distance.h
@@ -9,13 +9,15 @@
 namespace distance_metrics {
 
     // FIX17.md Step 6: Inlined distance metric functions for performance
+    // FIX22 Tier 1: Use std::sqrtf for float operations to avoid casting
+    // FIX22 Tier 2: Add const pointers for better compiler optimization
     inline float euclidean_distance(const float* a, const float* b, int dim) {
         float dist = 0.0f;
         for (int i = 0; i < dim; ++i) {
             float diff = a[i] - b[i];
             dist += diff * diff;
         }
-        return std::sqrt(dist);
+        return std::sqrtf(dist);
     }
 
     inline float cosine_distance(const float* a, const float* b, int dim) {
@@ -27,8 +29,9 @@ namespace distance_metrics {
             norm_b += b[i] * b[i];
         }
 
-        norm_a = std::sqrt(norm_a);
-        norm_b = std::sqrt(norm_b);
+        // FIX22 Tier 1: Use std::sqrtf for float operations
+        norm_a = std::sqrtf(norm_a);
+        norm_b = std::sqrtf(norm_b);
 
         if (norm_a < 1e-10f || norm_b < 1e-10f) return 1.0f;
 
@@ -66,7 +69,8 @@ namespace distance_metrics {
 
         if (den_a < 1e-10f || den_b < 1e-10f) return 1.0f;
 
-        float correlation = num / std::sqrt(den_a * den_b);
+        // FIX22 Tier 1: Use std::sqrtf for float operations
+        float correlation = num / std::sqrtf(den_a * den_b);
         correlation = std::max(-1.0f, std::min(1.0f, correlation));
 
         return 1.0f - correlation;
@@ -157,6 +161,7 @@ namespace distance_metrics {
     }
 
     // Unified distance computation based on metric type
+    // FIX22 Tier 2: Add const pointers for better compiler optimization
     float compute_distance(const float* a, const float* b, int dim, PacMapMetric metric);
     double compute_distance(const double* a, const double* b, int dim, PacMapMetric metric);
